feat: splitted UOView in separated sections components
